Add a rasterizer rendering approach.  Make it the default one.
Make the draw distance 10 times larger when using the rasterizer (it
is still faster).  Make the flashlight brighter so the improved draw
distance is actually useful.

Also modify the culling feature slightly in the rasterizer compared
with the raycast approaches.  It's still not quite good.

The rasterizer also brings with it new, fun rendering artifacts.